requirements.txt
streamlit>=1.40.0
google-genai>=0.3.0
fastmcp>=0.4.1
pydantic>=2.0.0
python-dotenv>=1.0.0